빌드 오류 수정

+++ src/components/board/main/comment/BoardCommentItem.tsx
@@ -9,9 +9,13 @@ import {getReplyLikes, getUserIsReplyLikes, ILikesDto, registerLikes} from "src/
 import FavoriteBorderOutlinedIcon from '@material-ui/icons/FavoriteBorderOutlined';
 import FavoriteIcon from "@material-ui/icons/Favorite";
 import ReportGmailerrorredOutlinedIcon from '@material-ui/icons/ReportGmailerrorredOutlined';
+import {useRouter} from "next/router";
+import {getCookie} from "../../../../utils/cookieUtil";
 
 const BoardCommentItem = ({reply} : IReplyProps) => {
+    const userId = getCookie('userId');
     const queryClient = useQueryClient();
+    const router = useRouter();
     const [showAdd, setShowAdd] = useState(false);
 
     const replyLikesQuery = useQuery<ILikesDto>(['replyLikes',reply.id],() => getReplyLikes(reply.id), {
@@ -43,15 +47,21 @@ const BoardCommentItem = ({reply} : IReplyProps) => {
     }
 
     const onDeleteReply = async () => {
-        const response = await deletePostReply(reply.id);
-        console.log(response);
-    }
+        if(userId === reply.userId)
+        {
+            const response = await deletePostReply(reply.id);
+            console.log(response);
+        }
 
+        await queryClient.invalidateQueries('postReply');
+        await queryClient.invalidateQueries('replyQuery');
+    }
+``
     const totalReplyQuery = useQuery<number>(['totalReply', reply.id],() => getTotalReply(reply.id), {
         staleTime : 1000 * 60
     })
 
-    const replyQuery = useQuery<IReplyDto[]>(['replyQuery',reply.id], () => getReply(reply.id), {
+    const replyQuery = useQuery<IReplyDto[]>(['replyQuery', reply.id], () => getReply(reply.id), {
         staleTime : 1000 * 60
     })
 
@@ -86,6 +96,7 @@ const BoardCommentItem = ({reply} : IReplyProps) => {
                 </div>
                 {
                     replyQuery.data?.map((reply) => (
+                        !reply.isDeleted &&
                         <BoardCommentItem key = {reply.id} reply={reply}/>
                     ))
                 }
